dcs/requirements.txt
ipython==7.24.1
pandas==1.1.1
numpy==1.19.5
h5py==2.10.0
tables==3.6.1
PyYAML==5.4.1
scikit-learn==1.0.2
lifelines==0.25.4
pycox==0.2.1
# torch==1.6.0
scikit-survival==0.17.1